chromium/third_party/blink/web_tests/virtual/no-auto-wpt-origin-isolation/external/wpt/origin-isolation/getter-sandboxed-iframe.sub.https-expected.txt

This is a testharness.js-based test.
FAIL originIsolationRestricted must equal false assert_equals: expected false but got true
Harness: the test ran to completion.
